Added count of risk factors
Man, this took a lot of time and work...
Digging through the KB and texts sent to other people to try to achieve this.  I'd first like to think the MSAI program which has supported my studies for the better part of a day (in total).  Also, I'd like to apologize to my dog, because I haven't seen him in a while....WHo are you? Oh god, they're taking me away, se4nd hekp this is my final messsssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss